Day17...
This is hard-coded for my particular input, and is probably slower than it needs to be, but the puzzle's explaination for how it would prompt for input was all but non-existent and I had to go to Reddit and scroll down a hundred responses to find it. So this works and its good enough.