2021 Day 18
I'm not proud of this one. I was busy with other stuff today and didn't get a good time to completely understand the instructions until late in the evening. It works. Its reasonably fast. I might make a better solution some other time if I feel like coming back to it.